# server/docs/swagger.yaml
basePath: /
definitions:
...
  dto.ContentPermissions:
    properties:
      comment:
        type: boolean
      delete:
        type: boolean
      edit:
        type: boolean
      view:
        type: boolean
      viewBody:
        type: boolean
      vote:
        type: boolean
    required:
    - comment
    - delete
    - edit
    - view
    - viewBody
    - vote
    type: object
...
  dto.DiscussionResponse:
    properties:
      authorId:
        type: string
      authorName:
        type: string
      contentMd:
        type: string
      createdAt:
        type: string
      domainId:
        type: string
      edited:
        description: Edited is true when the post was changed after it was written.
        type: boolean
      editorialId:
        type: string
      id:
        type: integer
      parentId:
        type: integer
      permissions:
        $ref: '#/definitions/dto.ContentPermissions'
      problemId:
        type: string
      updatedAt:
        type: string
    required:
    - contentMd
    - createdAt
    - domainId
    - edited
    - id
    - permissions
    - updatedAt
    type: object
  dto.DiscussionThreadResponse:
    properties:
      canPost:
        type: boolean
      items:
        items:
          $ref: '#/definitions/dto.DiscussionResponse'
        type: array
      total:
        type: integer
    required:
    - canPost
    - items
    - total
    type: object
  dto.DiscussionUpdateRequest:
    properties:
      contentMd:
        type: string
    required:
    - contentMd
    type: object
...
  dto.EditorialDiscussionCreateRequest:
    properties:
      contentMd:
        type: string
      parentId:
        type: integer
    required:
    - contentMd
    type: object
...
  dto.ProblemDiscussionCreateRequest:
    properties:
      contentMd:
        type: string
      parentId:
        type: integer
    required:
    - contentMd
    type: object
...
  httpx.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
    required:
    - code
    - error
    type: object
...
  httpx.StatusResponse:
    properties:
      status:
        type: string
    required:
    - status
    type: object
info:
  contact: {}
  description: Public application endpoints and the authenticated Judge worker protocol.
  title: Vertex OJ API
  version: "1.0"
paths:
...
  /api/discussions/{postId}:
    delete:
      parameters:
      - description: Discussion ID
        in: path
        name: postId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpx.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete discussion
      tags:
      - discussions
    put:
      consumes:
      - application/json
      parameters:
      - description: Discussion ID
        in: path
        name: postId
        required: true
        type: integer
      - description: New content
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.DiscussionUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DiscussionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Edit a discussion post
      tags:
      - discussions
...
  /api/editorials/{id}/discussions:
    get:
      parameters:
      - description: Editorial ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DiscussionThreadResponse'
      summary: List editorial discussions
      tags:
      - discussions
    post:
      consumes:
      - application/json
      parameters:
      - description: Editorial ID
        in: path
        name: id
        required: true
        type: string
      - description: Discussion
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.EditorialDiscussionCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.DiscussionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create editorial discussion
      tags:
      - discussions
...
  /api/problems/{id}/discussions:
    get:
      parameters:
      - description: Problem ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DiscussionThreadResponse'
      summary: List problem discussions
      tags:
      - discussions
    post:
      consumes:
      - application/json
      parameters:
      - description: Problem ID
        in: path
        name: id
        required: true
        type: string
      - description: Discussion
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ProblemDiscussionCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.DiscussionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create problem discussion
      tags:
      - discussions
...
securityDefinitions:
  BearerAuth:
    description: Prefix the access token with "Bearer ".
    in: header
    name: Authorization
    type: apiKey
  JudgeServiceAuth:
    description: Internal Judge service credential using the Bearer scheme.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"